backend/app/requirements.txt
fastapi
uvicorn[standard] 
python-dotenv     
faiss-cpu         
sentence-transformers
numpy
google-generativeai
pydantic          
sqlite3           